rename mb to mobilenet
