+++ configs/nas/spos/spos_evolution_search_mobilenet_8xb1024_in1k.py
@@ -1,4 +1,4 @@
-_base_ = ['./spos_supernet_mb_8xb128_in1k.py']
+_base_ = ['./spos_supernet_mobilenet_8xb128_in1k.py']
 
 data = dict(
     samples_per_gpu=1024,
