+++ configs/nas/spos/spos_mobilenet_for_check_ckpt_from_anglenas.py
@@ -1,5 +1,5 @@
 _base_ = [
-    './spos_subnet_mb_proxyless_gpu_8xb128_in1k.py',
+    './spos_subnet_mobilenet_proxyless_gpu_8xb128_in1k.py',
 ]
 
 img_norm_cfg = dict(mean=[0., 0., 0.], std=[1., 1., 1.], to_rgb=False)
