+++ configs/nas/spos/spos_supernet_mobilenet_8xb128_in1k.py
