+++ mmrazor/models/ops/__init__.py
@@ -2,7 +2,7 @@
 from .common import Identity
 from .darts_series import (DartsDilConv, DartsPoolBN, DartsSepConv,
                            DartsSkipConnect, DartsZero)
-from .mb_series import MBBlock
+from .mobilenet_series import MBBlock
 from .shufflenet_series import ShuffleBlock, ShuffleXception
 
 __all__ = [
